+++ CHANGELOG.md
@@ -18,6 +18,7 @@ ENHANCEMENTS:
 
 BUG FIXES:
 
+- The `-json` apply output now emits the periodic `apply_progress` messages for resources that take more than 10 seconds, instead of only the first one per resource. ([#4107](https://github.com/opentofu/opentofu/issues/4107))
 - The built-in function `contains` now accepts `null` as its second argument, to test whether a collection contains any null values. ([#4043](https://github.com/opentofu/opentofu/issues/4043))
 - The built-in function `merge` no longer fails when its only argument is a null value of an object type. ([#4043](https://github.com/opentofu/opentofu/issues/4043))
 - The built-in function `cidrhost` no longer returns a "panic" error when called with an out-of-range host number represented in more than 64 bits. ([#4056](https://github.com/opentofu/opentofu/pull/4056))